Make redirect behaviour match with expectations of RFC 3986
Added a new interface for upstream elements to be able to query for
a redirected URI. This is useful for DASH streams where the MPD is
being redirected, and for MPDs without a BaseURL attribute all media
segments will take the MPD base URL and construct relative paths from
it. We must know the base URL of the actual manifest file in order for
those relative paths to work properly.

There is a known flaw where there is currently no way via the libcurl
interface to tell if the redirect is temporary or permanent. Hopefully at
some point in the future I will find the time to read the headers and
set a flag dependant on the specific 3xx returned by the redirector.