Fix GLib runtime warnings
There was an issue when adaptivedemux/dashdemux would call to set the
extra_headers property but there were no headers to set, thereby causing
a NULL pointer to be passed to g_strv_length(). This meant a
GLib-CRITICAL warning was issued to the command line. This simple fix
checks for a NULL pointer first, and if it's invalid then just set the number
of headers as 0 to prevent the issue.